fix updateItem exports
